[[TEST]] Re-factor tests for "module" API
Separate the project's single unit test for the module API into a
standalone file. Beyond improving test discoverability, this
organization encapsulates new logic necessary for test hygiene (as
described in the in-line documentation).
